# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "302":
          description: Redirect to the OIDC provider logout URL.
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, isAdmin]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
          format: email
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
        isAdmin:
          type: boolean
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
...